Created `include` mustache tag
created an include tag that simply gets replaced with the contents of a
file. This is a quick and dirty patch so I can incorporate the html
files output from --describeTags into the documentation. I've only
updated one of the files tags-abstract. I don't really like how it looks
but I can fix that later. The important part is that it's relatively
easy to include this information now.