Pulled content for articles into extended_docs
For articles other than tags, I pulled their content to the
Jake\extended_docs\ folders and put them into subfolders using their
filename prefix. The contents of the files is included into the
Jake\articles file so everything looks the same. The intent of this is
to be able to map files into articles from various sources and maintain
a single directory for user contributed files. As it stands, any files
in the extended_docs folder would be hand written in HTML with the only
special requirement being that code examples use the mustache example
tag if they want to utilize whatever source highlighter we're using.