Updated readme
kastor.js was a jake file I used to sort lots of data without ruining my
wrists and fingers. It is no longer necessary. articlePrep.bat is still
useful because it will automatically regenerate the html files from
jsdoc --describeTags and then rebuild the documentation project. I want
to incorporate this action into the jake file in the future.