Now automatically detects new tags
When I generate html files using --describeTags and then incorporate
them into the build process, I can use them to do directory comparisons
and figure out if an article file or extended doc is missing. If the
article file for a tag is missing then it will be generated
automatically and have the proper content to link it to the extended
docs and definition. A placeholder extended doc will also be generated
although it doesn't say much.

see many many tags that were undocumented before (15),
generateMissingTagsArticles.js, and phpjs's node module.