updated build system.
Using current head version of mustache because the old one had bugs.
Using phpjs because it's awesome.
Using colors because it just looks better in crayon.

Still don't know how to generate good docs for the JSDoc 3 API. . . If
only the source code were in markdown... or fucking YAML! :O Imagine
that, if we could write programs in prose...