Examples can now be includes, will be prettified.
new mustache tag, includeExample will fetch the contents of a file and
insert it into the document. The filename will be used as the title
above it's display. The example tag still works but this is much easier
to mess with: just drop a javascript file in the examples folder and put
the includeExample tag where you want it to show up. You don't have to
worry about special characters or any other silly b.s.